// src/models/WeatherStateHourly.ts
export interface WeatherStateHourly {
  cloudcover: number
  conditions: string
  datetime: string
  datetimeEpoch: number
  dew: number
  feelslike: number
  humidity: number
  icon: string
  precip: number
  precipprob: number
  preciptype?: number
  pressure: number
  severerisk: number
  snow: number
  snowdepth: number
  solarenergy: number
  solarradiation: number
  temp: number
  uvindex: number
  visibility: number
  winddir: number
  windgust: number
  windspeed: number
}